packages/vector-db/pyproject.toml
[tool.poetry]
name = "vector-db"
version = "0.1.0"
description = ""
authors = ["Mark Shteyn <markshteyn1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
weaviate-client = "^4.9.6"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
